requirements.txt
aiogram==3.0.0b8
aiohttp>=3.8.6
Pillow>=10.3.0
pydantic>=2.0.3
pydantic_settings>=2.0.3
redis>=5.0.6